feat: add DataQualityOperator and adding a DataQualityOperator-based task to the DAG, add DAG configurations

+++ plugins/operators/load_dimension.py
@@ -11,17 +11,22 @@ class LoadDimensionOperator(BaseOperator):
     def __init__(self,
                  redshift_conn_id="",
                  table="",
+                 append_only=False,
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
         self.redshift_conn_id = redshift_conn_id
         self.table = table
+        self.append_only = append_only
 
     def execute(self, context):
         redshift = PostgresHook(self.redshift_conn_id)
 
-        self.log.info('Remove pre-existing data...')
-        redshift.run(f'DELETE FROM {self.table}')
+        if self.append_only:
+            self.log.info('Remove pre-existing data...')
+            redshift.run(f'DELETE FROM {self.table}')
+        else:
+            self.log.info('Pre-existing data is preseved... new data is appended only.')
 
         sql_query = insert_table_queries[self.table]
         self.log.info(f"Load data to table {self.table} executing {sql_query} ...")
